extract fake state machine.
