+++ src/raft/mod.rs
@@ -22,3 +22,6 @@ mod diagnostics;
 mod log;
 mod state_machine;
 mod store;
+
+#[cfg(test)]
+mod testing;
